requirements.txt
langchain==0.1.17
langchain-core==0.1.52
langchain-community==0.0.36
langchain-groq
langchain-huggingface
langchain-chroma
langchain-openai
langsmith
chromadb
sentence-transformers
faiss-cpu
streamlit
pypdf
pymupdf
unstructured
python-dotenv
pandas
duckdb
numexpr
wikipedia
duckduckgo-search
arxiv
youtube_transcript_api
pytube
mysql-connector-python
SQLAlchemy
validators==0.28.1
torchvision
ipykernel
openai